Restructure file directory
